// types/starships.ts
export interface Starships {
   cargo_capacity: string;
   consumables: string;
   cost_in_credits: string;
   created: Date;
   crew: string;
   edited: Date;
   films: string[];
   hyperdrive_rating: string;
   length: string;
   manufacturer: string;
   max_atmosphering_speed: string;
   MGLT: string;
   model: string;
   name: string;
   passengers: string;
   pilots: string[];
   starship_class: string;
   url: string;
}